requirements.txt
openai==0.27.0
discord.py==1.7.3
python-dotenv==0.19.1